Reintroduce multi inputs/outputs with new types
For simplicity it was decided that a configuration file should only
require a single input and output. However, there is still a valid use
case for having multiple options for each at run time. I've started
reintroducing these concepts as a separate input and output type (fan_in
and fan_out).

The new types have an array of inputs or outputs, allowing you to add
complexity as required. This also makes it easier for me to introduce
new patterns such as round-robin or random selection as new output
types.