Add env var injection for streams API input
The standard Benthos runtime allows for the use of "${}" as a
placeholder for environment variables and injects the value of those
variables in to a configuration at startup time. This patch adds the
same behavior to the YAML/JSON documents given to the streams API
endpoints when creating or updating a stream.