Add support for SASL access token providers to Kafka input/outputs
This adds support for Sarama's SASL access token providers to the Kafka
inputs and outputs. Since access token providers are callbacks a custom
build of Benthos that registers the access token provider under a given
name must be used; that name is then used in the config to specify
which access token provider to use.

A static access token can also be specified via `access_token`, this is
useful for testing.