Fix schema_registry_decode processor
When dealing with logical types, we can't serialise the native
struct emitted by goavro directly to JSON, since that will discard
schema information that `codec.TextualFromNative` uses to produce
the expected JSON.

Also, the tip version of goavro is required because the fix for
this regression https://github.com/linkedin/goavro/issues/242 was
merged, but there isn't a tagged release yet.

Fixes #1161.