feat: allow delaying shutdown of benthos process
This change adds a top-level `shutdown_delay` config field that sets an optional shutdown delay duration. This feature is useful for pipelines that run for a finite number of messages and terminate but need a short time window to have their final metrics and traces scraped or pushed out.

The shutdown delay is triggered when the pipeline terminates after running out of new messages to process. Other shutdown triggers such as receiving SIGTERM, closing the HTTP server or a propagating context cancellation do not trigger this shutdown delay. This is because those triggers are considered explicit or catastrophic. For example, if the http server is closed then metrics cannot be scraped and delaying the shutdown procedures is not useful. If receiving an OS signal to terminate, then we assume this is an explicit request to shutdown and we shouldn't delay it.