Fix aws_sqs backoff reset behavior
The current implementation results in artificial delays when reading
from a queue that has an average receive message frequency lower than
the configured receive wait timeout. For example, if a queue receives,
on average, a burst of messages every 15-30m, and this input is
configured with long polling enabled (using the max receive wait timeout
of 20s), most messages will experience an artificial delay of 5m. This
change removes the conditional clause around receiving > 0 messages for
resetting the backoff, and instead resets it on every successful poll.