Optimize cache hot-path in dynamo, s3, ristretto, memcache, and redis.
Each time cache operations are invoked we pool.Get() a
backoff.BackOff and use a defer func to reset the
BackOff and then place it back in the pool.

Currently the BackOff is only necessary if a cache operation fails and
the backoff needs to wait and attempt a retry.  This commit moves the
Pool.Get() and the defer to a conditional to the failure path.

This completely eliminates the need for the defer that performs the Pool Get/Put and the reset of the
BackOff object in success conditions.