More unit test fixes
It turns out the newer protobuf packages we've upgraded to are oddly inconsistent with how they format JSON data and error messages. This has resulted in some annoying heisenbugs. Sorry for anyone that got caught out by these in your PRs, the solution is to replace all the content equals test conditions for either contains or a json aware compare.

Signed-off-by: Tiago Peczenyj <tpeczenyj@weborama.com>