Add new grouping processor and its test in Benthos
This commit introduces a new grouping processor in the Benthos package for batching multiple sub-processors together under a single processor. This new processor will allow to retain message batches and improve the readability and testability of configurations. Additionally, a new unit test has been created to ensure the processor's functionality.

Signed-off-by: Daan Gerits <daan.gerits@gmail.com>