Add test runner implementation
This commit introduces a new test runner implementation. The runner has two kinds, "legacy" and "neo", which can be selected upon initialization. The test runner also provides a unified interface for running test configurations. This development will simplify the test running process and allow for further extensions in the future.

Signed-off-by: Daan Gerits <daan.gerits@gmail.com>