Add `keys_filter` field to nats_kv processor
This is a slightly-breaking change to the nats_kv processor which
is still beta. It alters the behaviour of the `keys` operation
which now returns an array of keys instead of a batch of messages,
as discussed with @codegangsta.

I'm a bit concerned that the history operation still returns a
batch of messages, which might confuse people. Happy to modify
that one as well to return an array of objects if you'd like. In
the current implementation, some of the fields of each history
item are emitted as message metadata, which is a bit odd anyway,
since they're not identical across the batch...

Signed-off-by: Mihai Todor <todormihai@gmail.com>