Fix CI docs check and do some cleanup
Currently, we don't detect if a new component got added without
having its docs committed to the repo. With `git diff-index` we
can only detect files which git is aware of. Some details here:
https://stackoverflow.com/questions/2657935/checking-for-a-dirty-index-or-untracked-files-with-git

Also clean up govulncheck, since Dependabot seems to be good
enough for this and golangci-lint doesn't want to add this linter
to their suite.

Signed-off-by: Mihai Todor <todormihai@gmail.com>