introduce ollama processor
The ollama_completion and ollama_embeddings processor can complete text
and generate embeddings respectively. The use ollama.com server to
perform these operations and with the x_benthos_extra flags will also
embed a server and launch it locally.

I tried to embed the ollama server library directly instead of embedding
the entire server, but it's not quite setup for that at the moment (go
generate is needed in the dependency and I couldn't find a good way to
do that).

Otherwise it's a fairly standard barebones processor, there are more
options for requests and models that can be exposed but I figured that I
would start with the MVP and iterate from there.