Support pgvector for postgres outputs
Introduce support for `pgvector` using `pgvector-go` in our SQL outputs.

This is an opt in conversion using the new `.vector` function. When a
postgres output sees this it will convert it to a `pgvector-go` Vector.
We use a generic `vector` function so that we can use it for other
drivers such as clickhouse later.

If we ever switch to using the `pgx` driver, then we need to make sure
we register the pgvector type via:

```golang
err := pgxvector.RegisterTypes(ctx, conn)
```

Adds test via the pgvector docker image.