snowpipe: use generic numeric column narrowing
Okay this is kind of a "draw the rest of the owl" commit. We move
to a slightly different model of constructing parquet files.

Now in our code while we're iterating over the user batch we construct
the columnar format for the parquet file, and during that we also write
to buffers of a specific type to remove some indirection, this allows us
to control the numeric type narrowing when constructing the actual
parquet files in a second phase.