snowpipe: remove SQL query from startup
We should try not to always run a SQL query everytime we startup for
cost reasons. Instead of running a query (which is likely flaky because
of identifier normalization anyways), just open the channel lazily and
catch the specific error for the table not existing, then create the
table and retry.