schema_registry_decode: use hamba/avro as an alternative impl
I went back and forth on patching goavro, but this just seems simpler
and is much faster too. The tree/schema walking is a little complex but
here we are. Hopefully we can eventually just remove the code for goavro
at some point, but we have other code we'd need to switch around.