pgcdc: fix snapshot consistency
The previous approach here was busted. The *ONLY* safe way to get a
snapshot and a consistent start LSN for said snapshot is by creating a
replication slot. Any other technique is wrong.

We also don't want to create the replication slot until we've finished
the snapshot phase so if something fails, we don't skip the snapshot
phase as if it completed fine. Our workaround is to do the following:

1. Create a TEMPORARY replication slot with exported snapshot
2. Stream said snapshot.
3. Copy the replication slot to preserve the LSN, but make it not
   temporary (this requires PG >= 12, but I wish there was a command to
   make a replication slot permanent).
4. Start streaming from the initial LSN.

This was the only thing that I could think of that allowed for not
skipping the snapshot phase on restart without external state storage.

As such we drop support for two unsupported postgres versions, which I
think is fine.