mongodb/cdc: checkpoint when completing snapshot
This is useful if there are no changes to the DB before the connector
restarts.

High level we do this by plumbing a wait group to all ack functions and
waiting for all snapshot documents to be ack'd before moving on to the
streaming phase (and saving a new kind of checkpoint that signifies the
checkpoint is completed). This does add a "pause" in the pipeline before
moving into the streaming phase, but I'm hoping that won't be an issue in
practice.