mongodb/cdc: simplify checkpointing
We don't need a stop the world pause after the snapshot phase, we can
just commit the resume token if there are no new batches AND there is
nothing in flight at the moment. This is needed for low frequency
collections are being watched without changes but we still need to
ratchet forward our oplog position (encoded in the resumeToken).

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>