mcp: support rp internal flags
this makes it possible to setup logging to RP topic.

one slightly ugly thing: the logging schema writes a field pipeline_id.
this means, we will write a pipeline_id, even if it will be effectively
a mcp-server-id. theoretically, pipeline id and mcp-server-id can clash,
but in practice it will probably never happen.

since we are looking into an overhaul of the logging, i suggest to just
do this for now, and then rework it entirely (might be logging per
pipeline/mcp, for example).