protobuf: add faster proto->message path
The existing conversion to message is very slow and forces a JSON parse
if utilized after in a mapping or an output that uses structure (ie.
Snowflake or SQL). Add a fast path that skips the intermediate JSON
representation. Note this path is *less* lossy in terms of type
information when not roundtripping through JSON. I think this is more
what folks expect from a type prospective.