kafka(franz_reader_ordered): lazy header setting
Do not put individual headers to metadata by default, to use them we need to manually include them in the output.
It's better by default do nothing, and avoid allocations, as the headers most likely will not be used.
This could break some workloads relying on kafka headers, for this reason in only targets FranzReaderOrdered users i.e. redpanda_input and migrator.