protobuf: make it fast 🚀 (#3704)
* protobuf: add faster proto->message path

The existing conversion to message is very slow and forces a JSON parse
if utilized after in a mapping or an output that uses structure (ie.
Snowflake or SQL). Add a fast path that skips the intermediate JSON
representation. Note this path is *less* lossy in terms of type
information when not roundtripping through JSON. I think this is more
what folks expect from a type prospective.

* protobuf: split parsing utilities

I'm introducing a new parser for protobuf, which requires a different
form for protos. This means it's helpful to break down the methods to
parse and load protobuf into it's primative forms.

* protobuf: move parse utilities to common package

This will be needed in multiple places later

* deps: add hyperpb

* protobuf: introduce hyperpb parsing

Benchmark summary for this and the previous changes:
- hyperpb is ~10-40% faster than dynamicpb for decoding
- Fast conversion saves ~30% compared to slow conversion
- Reading after fast conversion has negligible overhead (data already structured)
- Reading after slow conversion adds ~60%+ overhead (requires JSON deserialization)

* protobuf: fixup parsing utilities

This corrects an issue in a previous commit where we dropped well-known
types when extracting descriptors

* protobuf: reduce allocations for hyperpb

* protobuf: enable PGO

To make hyperpb really fly

* protobuf: simplify decoder and plug into processors

This commit simplifies the common.NewDecoder method by just directly
taking a message descriptor, and then plugs it into the protobuf decoder
and the schema register decoder for protobuf.

This is up for discussion because the output will subtly change for the
protobuf decoders (for the better IMO).

* chore: make linters happy

* update changelog

* proto: make hyperpb not a breaking change

Make faster deserialization opt-in

* update changelog

* protobuf: allocate better